executer 管理提出来，准备集中精力搞一下
